Refactored into C++ class
